+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/aio/__init__.py
@@ -7,12 +7,12 @@
 from ._file_client_async import FileClient
 from ._directory_client_async import DirectoryClient
 from ._share_client_async import ShareClient
-from ._file_service_client_async import FileServiceClient
+from ._file_service_client_async import ShareServiceClient
 
 
 __all__ = [
     'FileClient',
     'DirectoryClient',
     'ShareClient',
-    'FileServiceClient',
+    'ShareServiceClient',
 ]