+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/aio/_file_service_client_async.py
@@ -21,7 +21,7 @@ from .._shared.policies_async import ExponentialRetry
 from .._generated.aio import AzureFileStorage
 from .._generated.models import StorageErrorException, StorageServiceProperties
 from .._generated.version import VERSION
-from .._file_service_client import FileServiceClient as FileServiceClientBase
+from .._file_service_client import ShareServiceClient as ShareServiceClientBase
 from ._share_client_async import ShareClient
 from ._models import SharePropertiesPaged
 
@@ -31,8 +31,8 @@ if TYPE_CHECKING:
     from .._models import Metrics, CorsRule, ShareProperties
 
 
-class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
-    """A client to interact with the File Service at the account level.
+class ShareServiceClient(AsyncStorageAccountHostsMixin, ShareServiceClientBase):
+    """A client to interact with the File Share Service at the account level.
 
     This client provides operations to retrieve and configure the account properties
     as well as list, create and delete shares within the account.
@@ -40,7 +40,7 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
     can also be retrieved using the `get_share_client` function.
 
     :ivar str url:
-        The full endpoint URL to the file service endpoint, including SAS token if used. This could be
+        The full endpoint URL to the file share service endpoint, including SAS token if used. This could be
         either the primary endpoint, or the secondard endpoint depending on the current `location_mode`.
     :ivar str primary_endpoint:
         The full primary endpoint URL.
@@ -58,7 +58,7 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
         The location mode that the client is currently using. By default
         this will be "primary". Options include "primary" and "secondary".
     :param str account_url:
-        The URL to the file storage account. Any other entities included
+        The URL to the file share storage account. Any other entities included
         in the URL path (e.g. share or file) will be discarded. This URL can be optionally
         authenticated with a SAS token.
     :param credential:
@@ -75,7 +75,7 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
             :end-before: [END create_file_service_client]
             :language: python
             :dedent: 8
-            :caption: Create the file service client with url and credential.
+            :caption: Create the share service client with url and credential.
     """
     def __init__(
             self, account_url,  # type: str
@@ -85,7 +85,7 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
         # type: (...) -> None
         kwargs['retry_policy'] = kwargs.get('retry_policy') or ExponentialRetry(**kwargs)
         loop = kwargs.pop('loop', None)
-        super(FileServiceClient, self).__init__(
+        super(ShareServiceClient, self).__init__(
             account_url,
             credential=credential,
             loop=loop,
@@ -96,12 +96,12 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
     @distributed_trace_async
     async def get_service_properties(self, **kwargs):
         # type: (Any) -> Dict[str, Any]
-        """Gets the properties of a storage account's File service, including
+        """Gets the properties of a storage account's File Share service, including
         Azure Storage Analytics.
 
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
-        :rtype: ~azure.storage.file._generated.models.StorageServiceProperties
+        :rtype: ~azure.storage.fileshare._generated.models.StorageServiceProperties
 
         .. admonition:: Example:
 
@@ -110,7 +110,7 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
                 :end-before: [END get_service_properties]
                 :language: python
                 :dedent: 8
-                :caption: Get file service properties.
+                :caption: Get file share service properties.
         """
         timeout = kwargs.pop('timeout', None)
         try:
@@ -126,23 +126,23 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
             **kwargs
         ):
         # type: (...) -> None
-        """Sets the properties of a storage account's File service, including
+        """Sets the properties of a storage account's File Share service, including
         Azure Storage Analytics. If an element (e.g. Logging) is left as None, the
         existing settings on the service for that functionality are preserved.
 
         :param hour_metrics:
             The hour metrics settings provide a summary of request
             statistics grouped by API in hourly aggregates for files.
-        :type hour_metrics: ~azure.storage.file.Metrics
+        :type hour_metrics: ~azure.storage.fileshare.Metrics
         :param minute_metrics:
             The minute metrics settings provide request statistics
             for each minute for files.
-        :type minute_metrics: ~azure.storage.file.Metrics
+        :type minute_metrics: ~azure.storage.fileshare.Metrics
         :param cors:
             You can include up to five CorsRule elements in the
             list. If an empty list is specified, all CORS rules will be deleted,
             and CORS will be disabled for the service.
-        :type cors: list(:class:`~azure.storage.file.CorsRule`)
+        :type cors: list(:class:`~azure.storage.fileshare.CorsRule`)
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
         :rtype: None
@@ -154,7 +154,7 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
                 :end-before: [END set_service_properties]
                 :language: python
                 :dedent: 8
-                :caption: Sets file service properties.
+                :caption: Sets file share service properties.
         """
         timeout = kwargs.pop('timeout', None)
         props = StorageServiceProperties(
@@ -188,7 +188,7 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
         :returns: An iterable (auto-paging) of ShareProperties.
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.storage.file.ShareProperties]
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.storage.fileshare.ShareProperties]
 
         .. admonition:: Example:
 
@@ -197,7 +197,7 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
                 :end-before: [END fsc_list_shares]
                 :language: python
                 :dedent: 12
-                :caption: List shares in the file service.
+                :caption: List shares in the file share service.
         """
         timeout = kwargs.pop('timeout', None)
         include = []
@@ -234,7 +234,7 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
             Quota in bytes.
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
-        :rtype: ~azure.storage.file.aio.ShareClient
+        :rtype: ~azure.storage.fileshare.aio.ShareClient
 
         .. admonition:: Example:
 
@@ -243,7 +243,7 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
                 :end-before: [END fsc_create_shares]
                 :language: python
                 :dedent: 8
-                :caption: Create a share in the file service.
+                :caption: Create a share in the file share service.
         """
         metadata = kwargs.pop('metadata', None)
         quota = kwargs.pop('quota', None)
@@ -266,7 +266,7 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
         :param share_name:
             The share to delete. This can either be the name of the share,
             or an instance of ShareProperties.
-        :type share_name: str or ~azure.storage.file.ShareProperties
+        :type share_name: str or ~azure.storage.fileshare.ShareProperties
         :param bool delete_snapshots:
             Indicates if snapshots are to be deleted.
         :keyword int timeout:
@@ -280,7 +280,7 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
                 :end-before: [END fsc_delete_shares]
                 :language: python
                 :dedent: 12
-                :caption: Delete a share in the file service.
+                :caption: Delete a share in the file share service.
         """
         timeout = kwargs.pop('timeout', None)
         share = self.get_share_client(share_name)
@@ -296,11 +296,11 @@ class FileServiceClient(AsyncStorageAccountHostsMixin, FileServiceClientBase):
         :param share:
             The share. This can either be the name of the share,
             or an instance of ShareProperties.
-        :type share: str or ~azure.storage.file.ShareProperties
+        :type share: str or ~azure.storage.fileshare.ShareProperties
         :param str snapshot:
             An optional share snapshot on which to operate.
         :returns: A ShareClient.
-        :rtype: ~azure.storage.file.aio.ShareClient
+        :rtype: ~azure.storage.fileshare.aio.ShareClient
 
         .. admonition:: Example: