+++ sdk/storage/azure-storage-file-share/tests/test_file_samples_authentication_async.py
@@ -31,10 +31,10 @@ class TestFileAuthSamples(FileTestCase):
     shared_access_key = settings.STORAGE_ACCOUNT_KEY
 
     async def _test_auth_connection_string(self):
-        # Instantiate the FileServiceClient from a connection string
+        # Instantiate the ShareServiceClient from a connection string
         # [START create_file_service_client_from_conn_string]
-        from azure.storage.file.aio import FileServiceClient
-        file_service = FileServiceClient.from_connection_string(self.connection_string)
+        from azure.storage.fileshare.aio import ShareServiceClient
+        file_service = ShareServiceClient.from_connection_string(self.connection_string)
         # [END create_file_service_client_from_conn_string]
 
         # Get queue service properties
@@ -48,13 +48,13 @@ class TestFileAuthSamples(FileTestCase):
         loop.run_until_complete(self._test_auth_connection_string())
 
     async def _test_auth_shared_key(self):
-        # Instantiate a FileServiceClient using a shared access key
+        # Instantiate a ShareServiceClient using a shared access key
         # [START create_file_service_client]
-        from azure.storage.file.aio import FileServiceClient
-        file_service_client = FileServiceClient(account_url=self.url, credential=self.shared_access_key)
+        from azure.storage.fileshare.aio import ShareServiceClient
+        file_service_client = ShareServiceClient(account_url=self.url, credential=self.shared_access_key)
         # [END create_file_service_client]
 
-        # Get account information for the File Service
+        # Get account information for the File Share Service
         account_info = await file_service_client.get_service_properties()
         assert account_info is not None
 
@@ -69,13 +69,13 @@ class TestFileAuthSamples(FileTestCase):
         if TestMode.need_recording_file(self.test_mode):
             return
 
-        # Instantiate a FileServiceClient using a connection string
-        from azure.storage.file.aio import FileServiceClient
-        file_service_client = FileServiceClient.from_connection_string(self.connection_string)
+        # Instantiate a ShareServiceClient using a connection string
+        from azure.storage.fileshare.aio import ShareServiceClient
+        file_service_client = ShareServiceClient.from_connection_string(self.connection_string)
 
         # Create a SAS token to use to authenticate a new client
         # [START generate_sas_token]
-        from azure.storage.file import generate_account_sas
+        from azure.storage.fileshare import generate_account_sas
 
         sas_token = generate_account_sas(
             file_service_client.account_name,