+++ sdk/storage/azure-storage-file-share/tests/test_file_samples_service.py
@@ -23,13 +23,13 @@ class TestFileServiceSamples(FileTestCase):
 
     @record
     def test_file_service_properties(self):
-        # Instantiate the FileServiceClient from a connection string
-        from azure.storage.file import FileServiceClient
-        file_service = FileServiceClient.from_connection_string(self.connection_string)
+        # Instantiate the ShareServiceClient from a connection string
+        from azure.storage.fileshare import ShareServiceClient
+        file_service = ShareServiceClient.from_connection_string(self.connection_string)
 
         # [START set_service_properties]
         # Create service properties
-        from azure.storage.file import Metrics, CorsRule, RetentionPolicy
+        from azure.storage.fileshare import Metrics, CorsRule, RetentionPolicy
 
         # Create metrics for requests statistics
         hour_metrics = Metrics(enabled=True, include_apis=True, retention_policy=RetentionPolicy(enabled=True, days=5))
@@ -62,9 +62,9 @@ class TestFileServiceSamples(FileTestCase):
 
     @record
     def test_share_operations(self):
-        # Instantiate the FileServiceClient from a connection string
-        from azure.storage.file import FileServiceClient
-        file_service = FileServiceClient.from_connection_string(self.connection_string)
+        # Instantiate the ShareServiceClient from a connection string
+        from azure.storage.fileshare import ShareServiceClient
+        file_service = ShareServiceClient.from_connection_string(self.connection_string)
 
         # [START fsc_create_shares]
         file_service.create_share(share_name="testshare")
@@ -87,8 +87,8 @@ class TestFileServiceSamples(FileTestCase):
     @record
     def test_get_share_client(self):
         # [START get_share_client]
-        from azure.storage.file import FileServiceClient
-        file_service = FileServiceClient.from_connection_string(self.connection_string)
+        from azure.storage.fileshare import ShareServiceClient
+        file_service = ShareServiceClient.from_connection_string(self.connection_string)
 
         # Get a share client to interact with a specific share
         share = file_service.get_share_client("fileshare")