+++ sdk/storage/azure-storage-file-share/tests/test_get_file.py
@@ -12,9 +12,9 @@ import unittest
 import pytest
 from azure.core.exceptions import HttpResponseError
 
-from azure.storage.file import (
+from azure.storage.fileshare import (
     FileClient,
-    FileServiceClient,
+    ShareServiceClient,
     FileProperties
 )
 from filetestcase import (
@@ -43,7 +43,7 @@ class StorageGetFileTest(FileTestCase):
         url = self.get_file_url()
         credential = self.get_shared_key_credential()
 
-        self.fsc = FileServiceClient(
+        self.fsc = ShareServiceClient(
             url, credential=credential,
             max_single_get_size=self.MAX_SINGLE_GET_SIZE,
             max_chunk_get_size=self.MAX_CHUNK_GET_SIZE)